move [TestNetwork] module
