+++ code/system/sys.net.webrtc/src/WebRtc.TestNetwork/common.ts
@@ -1,0 +1,12 @@
+export * from '../common';
+
+/**
+ * Test Constants
+ */
+export const DEFAULTS = {
+  signal: {
+    host: 'rtc.bus.events',
+    path: 'signal',
+    key: 'cell',
+  },
+} as const;
