+++ code/system/sys.net.webrtc/src/WebRtc.TestNetwork/types.mts
@@ -1,0 +1,7 @@
+import { type t } from './common';
+
+export type TestNetworkP2P = t.Disposable & {
+  peerA: t.Peer;
+  peerB: t.Peer;
+  connect(kind?: t.PeerConnectionKind[]): Promise<void>;
+};
