+++ code/system/sys.net.webrtc/src/WebRtc/-dev/-SPEC.tsx
@@ -13,7 +13,7 @@ import {
   rx,
   slug,
   type t,
-  TEST,
+  TEST_DEFAULTS,
   TextInput,
 } from '../../test.ui';
 import { PeerCard, PeerList } from '../../ui';
@@ -99,7 +99,7 @@ export default Dev.describe('WebRtc', async (e) => {
      * WebRTC (network).
      */
     const { getStream } = media;
-    self = await WebRtc.peer(TEST.signal, { getStream });
+    self = await WebRtc.peer(TEST_DEFAULTS.signal, { getStream });
     await state.change((d) => (d.self = self));
     self.connections$.subscribe((e) => state.change((d) => (d.self = self)));
   });
