+++ code/system/sys.net.webrtc/src/test.ui/common.ts
@@ -1,12 +1,1 @@
 export * from '../ui/common';
-
-/**
- * Test Constants
- */
-export const TEST = {
-  signal: {
-    host: 'rtc.bus.events',
-    path: 'signal',
-    key: 'cell',
-  },
-} as const;
