+++ code/system/sys.net.webrtc/src/test.ui/index.ts
@@ -5,6 +5,4 @@ export { TestFilesystem } from 'sys.fs.indexeddb';
 export { WebRtc } from '../index.mjs';
 export { WebRtcInfo } from '../ui/ui.Info';
 export { Mock } from './-dev.mocks';
-export { TestNetwork } from './TestNetwork.mjs';
-
-export type { TestNetworkP2P } from './TestNetwork.mjs';
+export { TestNetwork, DEFAULTS as TEST_DEFAULTS } from '../WebRtc.TestNetwork';
