+++ code/system/sys.net.webrtc/src/types.mts
@@ -1,6 +1,7 @@
 export type * from './sys.net.schema/types';
 
-export type * from './WebRtc/types.mjs';
 export type * from './WebRtc.Controller/types.mjs';
 export type * from './WebRtc.Events/types.mjs';
 export type * from './WebRtc.State/types.mjs';
+export type * from './WebRtc.TestNetwork/types.mjs';
+export type * from './WebRtc/types.mjs';
